requirements.txt
openai==0.27.0
pandas==2.0.3
python-dotenv==1.0.0
streamlit==1.26.0
